Renamed classes, updated some code
